Add local OAuth callback server to SwinubCLI
